src/requirements.txt
appnope==0.1.0
astroid==1.4.8
autopep8==1.2.4
bandit==1.1.0
decorator==4.0.10
Django==1.10.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-filter==0.15.3
djangorestframework==3.5.1
dodgy==0.1.9
flake8==3.0.4
gitdb2==2.0.0
GitPython==2.1.0
importmagic==0.1.7
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
jedi==0.9.0
lazy-object-proxy==1.2.2
mccabe==0.5.2
pbr==1.10.0
pep8==1.7.0
pep8-naming==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
prospector==0.12.4
ptyprocess==0.5.1
pycodestyle==2.0.0
pydocstyle==1.0.0
pyflakes==1.2.3
Pygments==2.1.3
pylint==1.6.4
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.4
PyYAML==3.12
requirements-detector==0.5.2
rope==0.10.3
setoptconf==0.2.0
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.1
sqlparse==0.2.2
stevedore==1.18.0
traitlets==4.3.1
uWSGI==2.0.14
wcwidth==0.1.7
wrapt==1.10.8
yapf==0.13.2
